ai-infra/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-infra"
version = "0.1.0"
description = "AI Infrastructure Generator — generates Docker, Compose, and Kubernetes configs from repository analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "typer>=0.12,<1",
    "rich>=13,<14",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "pyyaml>=6,<7",
    "jinja2>=3.1,<4",
    "fastapi>=0.111,<1",
    "uvicorn[standard]>=0.30,<1",
    "sse-starlette>=2,<3",
    "httpx>=0.27,<1",
    "gitpython>=3.1,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "coverage>=7,<8",
]

[project.scripts]
ai-infra = "cli.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["ai_infra", "cli"]
